+++ CHANGES.rst
@@ -1,9 +1,29 @@
 Changelog
 =========
 
+1.2 (unreleased)
+----------------
 
-1.0.dev1 (unreleased)
----------------------
+- incompatible changes (none yet)
+  [tobiasherp]
 
-- Initial release.
+
+1.1.2 (unreleased)
+------------------
+
+- interfaces module
+  [tobiasherp]
+
+
+1.1.1 (2018-09-26)
+------------------
+
+- New exception MissingFileError
+  [tobiasherp]
+
+
+1.1 (2018-09-17)
+----------------
+
+- First public release.
   [tobiasherp]